requirements.txt
fastapi
uvicorn[standard]
wsproto
numpy
pillow
rosidl-runtime-py
